backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
paho-mqtt==1.6.1
websockets==12.0
python-dotenv==1.0.0
pydantic==2.5.0
aiosqlite==0.19.0
